pyproject.toml
[project]
name = "flandre"
version = "0.1.0"
description = "芙兰天下第一可爱"
readme = "README.md"
requires-python = ">=3.8, <4.0"
dependencies = [
    "httpx>=0.28.1",
    "mcrcon>=0.7.0",
    "nonebot-adapter-onebot>=2.4.3",
    "nonebot-plugin-alconna>=0.42.4",
    "nonebot-plugin-apscheduler>=0.4.0",
    "nonebot-plugin-orm[sqlite]>=0.7.7",
    "nonebot2[fastapi]>=2.2.1",
    "pyotp>=2.9.0",
    "toml>=0.10.2",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
drivers = [
	{ name = "fastapi", module_name = "nonebot.drivers.fastapi" }
]
plugin_dirs = ["corePlugins"]